// server/enums/workPattern.ts
enum WorkPattern {
  ANNUALISED_HOURS = 'ANNUALISED_HOURS',
  COMPRESSED_HOURS = 'COMPRESSED_HOURS',
  FLEXI_TIME = 'FLEXI_TIME',
  FLEXIBLE_SHIFTS = 'FLEXIBLE_SHIFTS',
  JOB_SHARE = 'JOB_SHARE',
  STAGGERED_HOURS = 'STAGGERED_HOURS',
  TERM_TIME_HOURS = 'TERM_TIME_HOURS',
  UNSOCIABLE_HOURS = 'UNSOCIABLE_HOURS',
}

export default WorkPattern